Don't jinjify if there's no metadata
If a document comes in and there's no yaml frontmatter, we don't
jinjify.  You might not need any yaml frontmatter metadata but still
want jinja features (e.g. jinja macros).  In that case, add a dummy
var.

The reason for doing this is that legacy docs don't have any
frontmatter, and this is a convenient way to distinguish them.