Save last latex that builds
This might be useful for diffing against when your doc suddenly
doesn't build and you don't know why.  This doesn't save all the files
that go into making the doc.  It would be interesting to try to
snapshot them in some way.  It might even be interesting to collect
all the various files from all over and stik them in a dir so that
every build also makes a minimal one-dir version for slinging around.
Alas, this does none of that.