Load the 'changepage' package, for adjustwidth
Hmm, I'm a little leery of putting this here, given the advice in the
log message of commit 45c93c616d.  But we already have two docs
depending on it, and I didn't want to leave it sitting as a local
change (I'd meant to commit it and just forgot, and probably that all
happened before commit 45c93c616d was even made).  So I'm committing
it now, on the theory that it's a found dependency, but if we want to
revise our philosophy to be closer to what commit 45c93c616d says, I'm
open to that discussion, and I realize it might result in the
reversion of this change too.