Restore section title size hierarchy in contracts
Before this change, a section would have a small (5pt) title size, but
then subsections within it would have a larger title size -- I don't
know exactly what size, because it was some LaTeX default, but it was
clearly larger than 5pt in any case.

Probably the reason we never noticed this before is that our contracts
tended not to have subsections.  But then we wrote one that included
an SOW in the contract, and the SOW's subsections looked all wonky:
their title lines were larger than those of the containing sections.
This fixes the problem; I know not whether it introduces new problems.