Say more in error message about handling bug case
Since commit 018499ec3b, when the bug-detection plugin is enabled
(e.g. "bugs: True" in the header) then on detecting a bug in the
document we stop the build, unless in draft mode ("draft: True").

There are thus three ways to fix such a build:

  1) Fix whatever the bug is in the document.
  2) Turn on draft mode.
  3) Turn off bug-detection mode.

Since our bug-detection module is not perfect, (1) is not always an
option because there might not actually be a bug in the document.
That's what happened in this case: the module's attempt to detect
proper formatting of footnotes by looking for the leading backslash
(e.g., "\footnote" instead of "footnote") cries a false positive if
the document uses "\begin{footnotesize} ... \end{footnotesize}" to set
the font size.

I'll improve that bug-detection case in a subsequent commit.  In the
meantime, this commit improves the error message to inform the user
about possibilities (2) and (3), which are not otherwise obvious.