Only create venv if we can't satisfy python dependencies
I don't actually use the venv, and I was tired of it littering.
Virtual environments are probably best left to the user to configure,
and not for us to force on the user.  It's still nice to have it as a
backup, but if the user has a way to satisfy those dependencies
without our intervention, don't intervene.

There are a few ways to supply the python submodules that aren't in a
local venv.  One might install systemwide, use direnv, or use a
centralized venv for all of ots, etc.  I use direnv, but this change
makes room for any scheme a user might want.  We force our own venv
only as a last resort.