Fix order and conditionality of OTSDIR check
In https://github.com/OpenTechStrategies/ots-doctools/pull/10,
@choeppler points out (#pullrequestreview-425924575) that the "OTSDIR"
environment variable should not be required to be set, since only
"OTS_DOCTOOLS_DIR" is documented, and that furthermore if we are going
to check paths under "OTSDIR" we should at least do it in the same
order relative to "OTS_DOCTOOLS_DIR" that 'Makefile.docspecific' uses.

This commit is his suggested patch from that PR review.  (Note that
the commits in that PR are actually already applied.  There must have
been a rebasing somewhere along the way, as they all have different
commit IDs in master than they do in the PR.)