Use 'enumitem' package instead of 'enumerate'
The former supersedes the latter, and has more options, so we should
use it.  But the proxemitae cause for this change is that I wanted to
use lower-case Roman letters as item markers instead of numbers.  Now,
you might think it would work to just put "\usepackage{enumitem}" at
the top of the doc and then use "\begin{enumerate}[label=\alph*)]",
but no, that gets an error:

  ! Missing number, treated as zero.
  <to be read again>
                     \c@*
  l.394 \begin{enumerate}[label=\alph*)]

  !  ==> Fatal error occurred, no output PDF file produced!

The reason for this is that the two packages are incompatible.  With
'enumerate' already loaded, if you load 'enumitem' and then try to
take advantage of the feature for which you brought in 'enumitem' in
the first place, you'll get the above error.

This change should not break any of our older documents, since both
packages work fine for plain old enumerations or lists, and until now
that's all we could do -- we obviously weren't using the list-item
styling feature that 'enumitem' makes possible.  I also spot-checked
several of OTS's more complex documents to make sure they still build.