Further formatting improvements to invoices
Make "hour" and "hours" align on the right, which visually implies a
vertical line between the hours column and the charges column.  This
in turn makes right-aligning the charges column even more attractive
than it already was, so do that too (though it was probably a good
idea before and we just hadn't gotten around to doing it until now).