Switch to not removing refs by default
Before this change, refs would be stripped from documents by default
(even documents, such as invoices, that don't normally have any
explicit YAML head).  After this change, refs are not stripped by
default, but if you put "remove-refs: TRUE" in the YAML head then they
will be, and you can optionally customize the regex with "ref-regex".

Note also that it's "ref-regex" now instead of "ref_regex".  This is
because some editors (cough Emacs cough) render unescaped underscores
in a jarring and distracting way in LaTeX documents, including even
underscores in YAML frontmatter.  A better fix would be to teach Emacs
how to recognize when it's in a non-LaTeX section of a mostly-LaTeX
document, but that's a bit more yak than I have time to shave today.