Add report.ltx
This is identical to ${OTS_DIR}/forms/latex/report.ltx (as of r22457),
except that here we drop the "MACROS" section from the end -- all it
contained was the "contact()" macro, which gave James's and Karl's
contact information, which is not something we want to ship here.

This commit may or may not cause a problem for OTS folks building
reports, depending on whether this report.ltx shadows the one in
${OTS_DIR}/forms/latex/report.ltx.  If it does, we'll fix that when we
find out about it.  Right now, Karl and Seth Schoen just needed to get
ots-doctools working for Seth on his machine, and the obstacle in
front of them was that report.ltx was alleged to be missing:

  jinja2.exceptions.TemplateNotFound: report.ltx